requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
requests
streamlit
pytesseract
pillow
python-multipart